+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 ### Changed
 
 - upgraded the Toolbox plugin API, dropping support for Toolbox versions older than 3.7.2
-- display live workspace start progress reported by the Coder CLI
+- display live workspace start output from the Coder CLI and use polled build states to show progress while updating,
+  restarting, or stopping workspaces
 
 ## 0.9.4 - 2026-08-26